Add basic data structures and algorithms
We use basic data structures, that do not hold any data, and algorithms
that operate on them. These basic data structures can then be added as
headers to data. The data's header is then used with the algorithms.
Some data specific algorithms are also defined.

Note: this list is not complete, nor are there guarantees that all
algorithms are implemented and tested properly.