Add compiler demo
The demo shows how the interpreter can easily be used as a compiler. By
making it so that the top definition's instructions emit some sort of
output, be it raw binary or, as in this case, assembler instructions,
the output can be used to construct a binary that does the computation
that'd be done by the interpreter.

This demo is overly simplistic, and doesn't show the true extent to
where this technique can be used:

1. Loops and branches are ignored. Branching instructions would need to
   split execution to all of their destinations. If a jump were to lead
   to an specification that had already been process, execution would
   have to end so that the "compiler" doesn't output duplicates or get
   stuck in a loop.
2. The compiled output can be in another language. As long as the proper
   starting instructions are given, they can be used together to form
   any other instructions using a great enough number of definitions. In
   the most extreme case, only an `emit0` and `emit1` instruction would
   be required.
3. Adding new compilation targets is as easy as writing a couple of
   instruction definitions from simpler ones, or making base ones in the
   interpreter. This can culminate by having a large library of
   instruction definitions that utilize the aforementioned `emit0` and
   `emit1` instructions.